packages/backend/pyproject.toml
[tool.poetry]
name = "pranely-backend"
version = "0.1.0"
description = "PRANELY Backend API"
authors = ["PRANELY Team"]

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = "^0.115.0"
prometheus-client = "^0.19.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.30.0"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
redis = "^5.2.0"
rq = "^1.16.2"
httpx = "^0.27.0"
argon2-cffi = "^23.1.0"
python-jose = "^3.3.0"
email-validator = "^2.1.0"
stripe = "^7.0.0"

[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require Docker/PostgreSQL tools)",
    "slow: marks tests as slow (may take > 10s)"
]
